Moved Manual Testing code to a separate file manualtest.js
